pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "topcoat"
dynamic = ["version"]
description = 'Post-Quantum Threshold Digital Signature'
readme = "README.md"
requires-python = ">=3.7"
license = "MIT"
keywords = []
authors = [
    { name = "Nikita Snetkov", email = "nikita.snetkov@cyber.ee" },
    { name = "Jelizaveta Vakarjuk", email = "jelizaveta.vakarjuk@cyber.ee" },
    { name = "Peeter Laud", email = "peeter.laud@cyber.ee" },
    { name = "Petr Muzikant", email = "petr.muzikant@cyber.ee" },
]
classifiers = []
dependencies = ["click", "pyaml"]

# [project.urls]
# Documentation = "https://github.com/unknown/topcoat#readme"
# Issues = "https://github.com/unknown/topcoat/issues"
# Source = "https://github.com/unknown/topcoat"

[tool.hatch.version]
path = "src/topcoat/__about__.py"

[project.scripts]
topcoat = "topcoat.cli:topcoat_main"
